余额格式化，避免科学计数

+++ app/src/main/java/com/gs/buluo/app/bean/RequestBodyBean/WithdrawRequestBody.java
@@ -6,5 +6,5 @@ package com.gs.buluo.app.bean.RequestBodyBean;
 
 public class WithdrawRequestBody {
     public String bankCardId;
-    public float amount;
+    public double amount;
 }
